Add workflow for getting bios and update firmware
With Dell Racadm Tool
