+++ lib/task-data/base-tasks/dell-racadm-get-bios-base.js
@@ -3,8 +3,8 @@
 'use strict';
 
 module.exports = {
-    friendlyName: 'Dell Racadm Set BIOS base',
-    injectableName: 'Task.Base.Dell.Racadm.SetBIOS',
+    friendlyName: 'Dell Racadm Get BIOS base',
+    injectableName: 'Task.Base.Dell.Racadm.GetBIOS',
     runJob: 'Job.Dell.RacadmTool',
     requiredOptions: [
         "action"
