+++ lib/task-data/base-tasks/dell-racadm-update-firmware-base.js
@@ -3,8 +3,8 @@
 'use strict';
 
 module.exports = {
-    friendlyName: 'Dell Racadm Set BIOS base',
-    injectableName: 'Task.Base.Dell.Racadm.SetBIOS',
+    friendlyName: 'Dell Racadm Update Firmware Base',
+    injectableName: 'Task.Base.Dell.Racadm.Update.Firmware',
     runJob: 'Job.Dell.RacadmTool',
     requiredOptions: [
         "action"
