+++ lib/task-data/tasks/dell-racadm-get-bios.js
@@ -3,14 +3,14 @@
 'use strict';
 
 module.exports = {
-    friendlyName: 'dell racadm set bios',
-    injectableName: 'Task.Dell.Racadm.SetBIOS',
-    implementsTask: 'Task.Base.Dell.Racadm.SetBIOS',
+    friendlyName: 'dell racadm get bios',
+    injectableName: 'Task.Dell.Racadm.GetBIOS',
+    implementsTask: 'Task.Base.Dell.Racadm.GetBIOS',
     options: {
         server_username: null,
         server_password: null,
         server_filePath: null,
-        action: 'SetBIOS'
+        action: 'GetBIOS'
     },
     properties: {}
 };
