+++ spec/lib/task-data/base-tasks/dell-racadm-update-firmware-base-spec.js
@@ -1,0 +1,18 @@
+// Copyright 2016, EMC, Inc.
+/* jshint node:true */
+
+'use strict';
+
+describe(require('path').basename(__filename), function () {
+    var base = require('./base-task-data-spec');
+
+    base.before(function (context) {
+        context.taskdefinition = helper.require(
+            '/lib/task-data/base-tasks/dell-racadm-update-firmware-base.js');
+    });
+
+    describe('task-data', function () {
+        base.examples();
+    });
+
+});
