+++ spec/lib/task-data/tasks/dell-racadm-get-bios-spec.js
@@ -1,0 +1,17 @@
+// Copyright 2016, EMC, Inc.
+/* jshint node:true */
+
+'use strict';
+
+describe(require('path').basename(__filename), function () {
+    var base = require('./base-tasks-spec');
+
+    base.before(function (context) {
+        context.taskdefinition = helper.require('/lib/task-data/tasks/dell-racadm-get-bios.js');
+    });
+
+    describe('task-data', function () {
+        base.examples();
+    });
+
+});
